# backend/api/openapi.yaml
openapi: 3.0.0
info:
  title: UEDigitalEntity API
  description: NBI for UE Digital Entity
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: /
paths:
  /requirments:
    get:
      summary: get device requirments
      description: |
        Get the requirments of the associated physical device
      operationId: getUERequirments
      responses:
        "200":
          description: The requested requirments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Requirments'
                x-content-type: application/json
        "404":
          description: No requirment found
      x-swagger-router-controller: Requirments
    post:
      summary: create device requirments
      description: |
        Create the requirments of the associated physical device
      operationId: createUERequirments
      responses:
        "200":
          description: Requirments sucessefully created
      x-swagger-router-controller: Requirments
    patch:
      summary: update device requirments
      description: |
        Update the requirments of the associated physical device
      operationId: updateUERequirments
      responses:
        "200":
          description: The requirments to be updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Requirments'
                x-content-type: application/json
      x-swagger-router-controller: Requirments
  /gnb-connection-state:
    get:
      summary: get the UE connection state with gNB
      description: |
        Get the connection state of the UE with gNB
      operationId: getGnbConnectionState
      responses:
        "200":
          description: "The connection information of the UE with gNB (connectionState, gNB id)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GnbConnectionState'
        default:
          description: Generic-error
      x-swagger-router-controller: ConnectionState
  /subscription-info:
    get:
      summary: get the user equipment subscription information
      description: |
        Get the the user equipment subscription information querying the 5GS Digital Entity
      operationId: getUESubscriptionInfo
      responses:
        "200":
          description: "The subscription information of the UE (such as PDU Sessions,\
            \ QoS flows...)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionData'
        default:
          description: Generic-error
      x-swagger-router-controller: Subscription
  /pdu-session:
    get:
      summary: get the user equipment established PDU Sessions
      description: |
        Get the user equipment established PDU Sessions
      operationId: getUEPDUSessions
      responses:
        "200":
          description: The established PDU Sessions informations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EstablishedSession'
                x-content-type: application/json
      x-swagger-router-controller: ConnectionState
    post:
      summary: create a new PDU Session
      description: |
        Create a new PDU Session for the UE.
      operationId: createPDUSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectedSession'
      responses:
        "200":
          description: Pdu Session sucessfully established
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  emergency:
                    type: boolean
                  id:
                    type: integer
        default:
          description: Generic-error
      x-swagger-router-controller: ConnectionState

components:
  schemas:

    Requirments:
      type: object
      properties:
        field1:
          type: integer
        field2:
          type: string
        field3:
          type: string

    SubscriptionData:
      type: object
      properties:
        subscribedUeAmbr:
          $ref: '#/components/schemas/AmbrRm'
        nssai:
          $ref: '#/components/schemas/Nssai'

    Nssai:
      required:
      - defaultSingleNssais
      type: object
      properties:
        defaultSingleNssais:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/EnhancedSnssai'
        singleNssais:
          minItems: 0
          type: array
          items:
            $ref: '#/components/schemas/EnhancedSnssai'
      nullable: true

    EnhancedSnssai:
      required:
      - sst
      type: object
      properties:
        sst:
          maximum: 255
          minimum: 0
          type: integer
        sd:
          pattern: "^[A-Fa-f0-9]{6}$"
          type: string
        dnnConfigurations:
          description: A map (list of key-value pairs where Dnn, or optionally the Wildcard DNN, serves as key) of DnnConfigurations
          type: object
          additionalProperties:
             $ref: '#/components/schemas/DnnConfiguration'

    DnnConfiguration:
      required:
      - pduSessionTypes
      - sscModes
      type: object
      properties:
        pduSessionTypes:
          $ref: '#/components/schemas/PduSessionTypes'
        sscModes:
          $ref: '#/components/schemas/SscModes'
        "5gQosProfile":
          $ref: '#/components/schemas/SubscribedDefaultQos'
        sessionAmbr:
          $ref: '#/components/schemas/Ambr'
        staticIpAddress:
          maxItems: 2
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/IpAddress'

    IpAddress:
      type: object
      oneOf:
      - required:
        - ipv4Addr
      - required:
        - ipv6Addr
      - required:
        - ipv6Prefix
      properties:
        ipv4Addr:
          $ref: '#/components/schemas/Ipv4Addr'
        ipv6Addr:
          $ref: '#/components/schemas/Ipv6Addr'
        ipv6Prefix:
          $ref: '#/components/schemas/Ipv6Prefix'

    PduSessionTypes:
      required:
      - defaultSessionType
      type: object
      properties:
        defaultSessionType:
          $ref: '#/components/schemas/PduSessionType'
        allowedSessionTypes:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/PduSessionType'

    SscModes:
      required:
      - defaultSscMode
      type: object
      properties:
        defaultSscMode:
          $ref: '#/components/schemas/SscMode'
        allowedSscModes:
          maxItems: 3
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/SscMode'

    CellConnectionStatus:
      anyOf:
      - type: string
        enum:
        - CM-CONNECTED
        - CM-IDLE

    GnbConnectionState:
      required:
      - status
      type: object
      properties:
        status:
          $ref: '#/components/schemas/CellConnectionStatus'
        camped-cell:
          type: string

    SelectedSession:
      required:
      - sst
      type: object
      properties:
        sst:
          maximum: 255
          minimum: 0
          type: integer
        sd:
          pattern: "^[A-Fa-f0-9]{6}$"
          type: string
        dnn:
          type: string
        pduSessionType:
          $ref: '#/components/schemas/PduSessionType'

    EstablishedSession:
      required:
      - ipAddress
      - dnn
      - sd
      - sst
      type: object
      properties:
        id:
          type: integer
        sst:
          maximum: 255
          minimum: 0
          type: integer
        sd:
          pattern: "^[A-Fa-f0-9]{6}$"
          type: string
        dnn:
          type: string
        pduSessionTypes:
          $ref: '#/components/schemas/PduSessionType'
        ipAddress:
          $ref: '#/components/schemas/IpAddress'
        emergency:
          type: boolean

    AmbrRm:
      anyOf:
      - $ref: '#/components/schemas/Ambr'
      - $ref: '#/components/schemas/NullValue'

    Ambr:
      required:
      - downlink
      - uplink
      type: object
      properties:
        uplink:
          $ref: '#/components/schemas/BitRate'
        downlink:
          $ref: '#/components/schemas/BitRate'

    BitRate:
      pattern: ^\d+(\.\d+)? (bps|Kbps|Mbps|Gbps|Tbps)$
      type: string

    NullValue:
      type: string
      enum:
      - null

    SubscribedDefaultQos:
      required:
      - 5qi
      - arp
      type: object
      properties:
        "5qi":
          $ref: '#/components/schemas/5Qi'
        arp:
          $ref: '#/components/schemas/Arp'
        priorityLevel:
          $ref: '#/components/schemas/5QiPriorityLevel'

    "5Qi":
      maximum: 255
      minimum: 0
      type: integer

    Arp:
      required:
      - preemptCap
      - preemptVuln
      - priorityLevel
      type: object
      properties:
        priorityLevel:
          $ref: '#/components/schemas/ArpPriorityLevel'
        preemptCap:
          $ref: '#/components/schemas/PreemptionCapability'
        preemptVuln:
          $ref: '#/components/schemas/PreemptionVulnerability'

    ArpPriorityLevel:
      maximum: 15
      minimum: 1
      type: integer
      description: nullable true shall not be used for this attribute
      #nullable: true

    PreemptionCapability:
      anyOf:
      - type: string
        enum:
        - NOT_PREEMPT
        - MAY_PREEMPT
      #- type: string

    PreemptionVulnerability:
      anyOf:
      - type: string
        enum:
        - NOT_PREEMPTABLE
        - PREEMPTABLE
      - type: string

    "5QiPriorityLevel":
      maximum: 127
      minimum: 1
      type: integer

    Ipv4Addr:
      pattern: "^(([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\\.){3}([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])$"
      type: string

    Ipv6Addr:
      type: string
      example: 2001:db8:85a3::8a2e:370:7334
      allOf:
      - pattern: "^((:|(0?|([1-9a-f][0-9a-f]{0,3}))):)((0?|([1-9a-f][0-9a-f]{0,3})):){0,6}(:|(0?|([1-9a-f][0-9a-f]{0,3})))$"
      - pattern: "^((([^:]+:){7}([^:]+))|((([^:]+:)*[^:]+)?::(([^:]+:)*[^:]+)?))$"

    Ipv6Prefix:
      type: string
      example: 2001:db8:abcd:12::0/64
      allOf:
      - pattern: "^((:|(0?|([1-9a-f][0-9a-f]{0,3}))):)((0?|([1-9a-f][0-9a-f]{0,3})):){0,6}(:|(0?|([1-9a-f][0-9a-f]{0,3})))(\\\
          /(([0-9])|([0-9]{2})|(1[0-1][0-9])|(12[0-8])))$"
      - pattern: "^((([^:]+:){7}([^:]+))|((([^:]+:)*[^:]+)?::(([^:]+:)*[^:]+)?))(\\\
          /.+)$"

    PduSessionType:
      anyOf:
      - type: string
        enum:
        - IPV4
        - IPV6
        - IPV4V6
        - UNSTRUCTURED
        - ETHERNET
      - type: string

    SscMode:
      anyOf:
      - type: string
        enum:
        - SSC_MODE_1
        - SSC_MODE_2
        - SSC_MODE_3
      - type: string