# physical_ue_proxy/server/models/ipv4_addr.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from server.models.base_model_ import Model
from server import util


class Ipv4Addr(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self):  # noqa: E501
        """Ipv4Addr - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'Ipv4Addr':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Ipv4Addr of this Ipv4Addr.  # noqa: E501
        :rtype: Ipv4Addr
        """
        return util.deserialize_model(dikt, cls)